% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_deviation.R
\name{CORT}
\alias{CORT}
\title{Calculate CORT}
\usage{
CORT(X, Y)
}
\arguments{
\item{X}{a vector, containing the data values of the first time series}

\item{Y}{a vector, of the same length as X, containing the data values
of the second time series}
}
\value{
an atomic vector, containing the CORT value [-1,1]
}
\description{
Calculates first order temporal correlation coefficient (CORT) between two
time series. This is based on equation 3 in Gao et al. (2019), which is in
turn from page 10 of Chouakria & Nagabhushan (2007)
}
\examples{
data("EuStockMarkets")
CORT(X=EuStockMarkets[ ,1], Y=EuStockMarkets[ ,2])

}
\references{
Gao, Y. et al. (2019) ‘Measuring spatio-temporal autocorrelation
in time series data of collective human mobility’, Geo-spatial Information
Science, 22(3), pp. 166–173. doi:10.1080/10095020.2019.1643609.
}
\author{
Nils Tjaden \href{mailto:n.b.tjaden@utwente.nl}{n.b.tjaden@utwente.nl}
}
