% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localmoran_st.R
\name{localmoran_st}
\alias{localmoran_st}
\title{Calculate spatio-temporal Moran's I}
\usage{
localmoran_st(
  x,
  listw,
  zero.policy = NULL,
  na.action = na.fail,
  conditional = TRUE,
  alternative = "two.sided",
  mlvar = TRUE,
  spChk = NULL,
  adjust.x = FALSE
)
}
\arguments{
\item{x}{a numeric matrix (or data.frame) of observations. Columns represent
space, rows represent time, so that each column contains a time series
measured at one location. The number of columns must equal the length of
the neighbors list in \code{listw}.}

\item{listw}{a \code{listw} object (see documentation for \code{spdep})}

\item{zero.policy}{default NULL, use global option value; if TRUE assign zero
to the lagged value of zones without neighbours, if FALSE assign NA}

\item{na.action}{a function (default \code{na.fail}), can also be na.omit or
na.exclude - in these cases the weights list will be subsetted to remove
NAs in the data. It may be necessary to set zero.policy to TRUE because
this subsetting may create no-neighbour observations. Note that only
weights lists created without using the glist argument to \code{nb2listw} may be
subsetted. If \code{na.pass} is used, zero is substituted for NA values in
calculating the spatial lag}

\item{conditional}{default TRUE: expectation and variance are calculated
using the conditional randomization null (Sokal 1998 Eqs. A7 & A8).
Elaboration of these changes available in Sauer et al. (2021). If FALSE:
expectation and variance are calculated using the total randomization null
(Sokal 1998 Eqs. A3 & A4).}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of greater, less or two.sided (default).}

\item{mlvar}{default TRUE: values of local Moran's I are reported using the
variance of the variable of interest (sum of squared deviances over n), but
can be reported as the sample variance, dividing by (n-1) instead; both are
used in other implementations.}

\item{spChk}{should the data vector names be checked against the spatial
objects for identity integrity, TRUE, or FALSE, default NULL to use
get.spChkOption()}

\item{adjust.x}{default FALSE, if TRUE, x values of observations with no
neighbours are omitted in the mean of x}
}
\value{
\code{Ii_st} local moran statistic

\code{E.Ii_st} expectation of local moran statistic

\code{Var.Ii_st} variance of local moran statistic

\code{Z.Ii_st} standard deviate of local moran statistic

\verb{Pr()_st} p-value of local moran statistic using pnorm()

In addition, an attribute data frame "quadr" indicating the
High-High etc. quadrants
}
\description{
Calculates a spatio-temporal version of local Moran's I (a.k.a
spatio-temporal LISA) as proposed by Gao et al. (2019).
}
\references{
Gao, Y. et al. (2019) ‘Measuring spatio-temporal autocorrelation
in time series data of collective human mobility’, Geo-spatial Information
Science, 22(3), pp. 166–173. doi:10.1080/10095020.2019.1643609.

Anselin, L. 1995. Local indicators of spatial association,
Geographical Analysis, 27, 93–115

Getis, A. and Ord, J. K. 1996 Local spatial statistics: an
overview. In P. Longley and M. Batty (eds) Spatial analysis: modelling in a
GIS environment (Cambridge: Geoinformation International), 261–277

Sokal, R. R, Oden, N. L. and Thomson, B. A. 1998. Local Spatial
Autocorrelation in a Biological Model. Geographical Analysis, 30. 331–354

Bivand RS, Wong DWS 2018 Comparing implementations of global and
local indicators of spatial association. TEST, 27(3), 716–748
doi:10.1007/s11749-018-0599-x

Sauer, J., Oshan, T. M., Rey, S., & Wolf, L. J. 2021. The
Importance of Null Hypotheses: Understanding Differences in Local Moran’s
under Heteroskedasticity. Geographical Analysis. doi:10.1111/gean.12304

Bivand, R. (2022), R Packages for Analyzing Spatial Data: A
Comparative Case Study with Areal Data. Geographical Analysis, 54(3),
488-518. doi:10.1111/gean.12319
}
\author{
Roger Bivand \href{mailto:Roger.Bivand@nhh.no}{Roger.Bivand@nhh.no} (original code of spdep::moran)

Jeff Sauer (conditional code)

Levi Wolf (conditional code)

Nils Tjaden \href{mailto:n.b.tjaden@utwente.nl}{n.b.tjaden@utwente.nl} (adjustments for spatio-temporal
use)
}
