% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_deviation.R
\name{magnitude_deviation}
\alias{magnitude_deviation}
\title{Calculate magnitude deviation}
\usage{
magnitude_deviation(X, Y)
}
\arguments{
\item{X}{a vector, containing the data values of the first time series}

\item{Y}{a vector, of the same length as X, containing the data values
of the second time series}
}
\value{
an atomic vector, containing the magnitude deviation
}
\description{
The wording in Gao et al. (2019) describing is slightly obscure. They make no
direct reference to any of the methods used in Chouakria & Nagabhushan
(2007), and indeed it appears to be a modification of the Euclidian distance
method that does not enforce positive values. The last paragraph of section
2.3 of Gao et al. (2019) explains that: "The quantity deviation determines
whether, and how much, the autocorrelation is positive or negative; then the
similarity of the temporal variance adjusts the degree of the correlation."
}
\examples{
data("EuStockMarkets")
magnitude_deviation(X=EuStockMarkets[ ,1], Y=EuStockMarkets[ ,2])

}
\references{
Gao, Y. et al. (2019) ‘Measuring spatio-temporal autocorrelation
in time series data of collective human mobility’, Geo-spatial Information
Science, 22(3), pp. 166–173. doi:10.1080/10095020.2019.1643609.
}
\author{
Nils Tjaden \href{mailto:n.b.tjaden@utwente.nl}{n.b.tjaden@utwente.nl}
}
