% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moran_st.R
\name{moran.test_st}
\alias{moran.test_st}
\title{Moran's I test for spatial autocorrelation (spatio-temporal)}
\usage{
moran.test_st(
  x,
  listw,
  randomisation = TRUE,
  zero.policy = NULL,
  alternative = "greater",
  rank = FALSE,
  na.action = na.fail,
  spChk = NULL,
  adjust.n = TRUE,
  drop.EI2 = FALSE
)
}
\arguments{
\item{x}{a numeric matrix (or data.frame) of observations. Columns represent
space, rows represent time, so that each column contains a time series
measured at one location. The number of columns must equal the length of
the neighbors list in \code{listw}.}

\item{listw}{a \code{listw} object (see documentation for \code{spdep})}

\item{randomisation}{variance of I calculated under the assumption of
randomisation, if FALSE normality}

\item{zero.policy}{default NULL, use global option value; if TRUE assign zero
to the lagged value of zones without neighbours, if FALSE assign NA}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of greater (default), less or two.sided.}

\item{rank}{logical value - default FALSE for continuous variables, if TRUE,
uses the adaptation of Moran's I for ranks suggested by Cliff and Ord
(1981, p. 46)}

\item{na.action}{a function (default \code{na.fail}), can also be na.omit or
na.exclude - in these cases the weights list will be subsetted to remove
NAs in the data. It may be necessary to set zero.policy to TRUE because
this subsetting may create no-neighbour observations. Note that only
weights lists created without using the glist argument to \code{nb2listw} may be
subsetted. If \code{na.pass} is used, zero is substituted for NA values in
calculating the spatial lag}

\item{spChk}{should the data vector names be checked against the spatial
objects for identity integrity, TRUE, or FALSE, default NULL to use
\code{get.spChkOption()}}

\item{adjust.n}{default TRUE, if FALSE the number of observations is not
adjusted for no-neighbour observations, if TRUE, the number of observations
is adjusted}

\item{drop.EI2}{default FALSE, if TRUE, emulate CrimeStat <= 4.02}
}
\value{
A list with class htest containing the following components:

\code{statistic} the value of the standard deviate of Moran's I.

\code{p.value} the p-value of the test.

\code{estimate}	the value of the observed Moran's I, its expectation and
variance under the method assumption.

\code{alternative} a character string describing the alternative
hypothesis.

\code{method}	a character string giving the assumption used for
calculating the standard deviate.

\code{data.name} a character string giving the name(s) of the data.
}
\description{
This is a 1:1 copy of \code{spdep::moran.test} with minor adjustments for use with
\code{moran_st}. See \code{?spdep::moran.test} for more information.
}
\references{
Cliff, A. D., Ord, J. K. 1981 Spatial processes, Pion
}
\author{
Roger Bivand \href{mailto:Roger.Bivand@nhh.no}{Roger.Bivand@nhh.no} (original code of
spdep::moran.test)

Nils Tjaden \href{mailto:n.b.tjaden@utwente.nl}{n.b.tjaden@utwente.nl} (minor adjustments for
spatio-temporal use)
}
