% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moran_st.R
\name{moran_st}
\alias{moran_st}
\title{Calculate spatio-temporal Moran's I}
\usage{
moran_st(x, listw, n, S0, zero.policy = NULL, NAOK = FALSE)
}
\arguments{
\item{x}{a numeric matrix (or data.frame) of observations. Columns represent
space, rows represent time, so that each column contains a time series
measured at one location. The number of columns must equal the length of
the neighbors list in \code{listw}.}

\item{listw}{a \code{listw} object (see documentation for \code{spdep})}

\item{n}{number of zones}

\item{S0}{global sum of weights}

\item{zero.policy}{default NULL, use global option value; if TRUE assign zero
to the lagged value of zones without neighbours, if FALSE assign NA}

\item{NAOK}{if 'TRUE' then any 'NA' or 'NaN' or 'Inf' values in x are passed
on to the foreign function. If 'FALSE', the presence of 'NA' or 'NaN' or
'Inf' values is regarded as an error.}
}
\value{
A list: \code{I} spatio-temporal Moran's I \code{K} sample kurtosis of x
}
\description{
Calculates a spatio-temporal version of (global) Moran's I as proposed by Gao
et al. (2019). Called by \code{moran.test_st}.
}
\references{
Gao, Y. et al. (2019) ‘Measuring spatio-temporal autocorrelation
in time series data of collective human mobility’, Geo-spatial Information
Science, 22(3), pp. 166–173. doi:10.1080/10095020.2019.1643609.
}
\author{
Roger Bivand \href{mailto:Roger.Bivand@nhh.no}{Roger.Bivand@nhh.no} (original code of spdep::moran)

Nils Tjaden \href{mailto:n.b.tjaden@utwente.nl}{n.b.tjaden@utwente.nl} (adjustments for spatio-temporal
use)
}
