% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_deviation.R
\name{ts_deviation}
\alias{ts_deviation}
\title{Calculate time series deviation}
\usage{
ts_deviation(X, Y)
}
\arguments{
\item{X}{a vector, containing the data values of the first time series}

\item{Y}{a vector, containing the data values of the second time series}
}
\value{
an atomic vector, containing the deviation between the two time
series
}
\description{
Measure for the deviation between two time series. This can be used to
estimate the difference between a specific time series and the mean time
series.
}
\examples{
data("EuStockMarkets")
ts_deviation(X=EuStockMarkets[ ,1], Y=EuStockMarkets[ ,2])

}
\references{
Gao, Y. et al. (2019) ‘Measuring spatio-temporal autocorrelation
in time series data of collective human mobility’, Geo-spatial Information
Science, 22(3), pp. 166–173. doi:10.1080/10095020.2019.1643609.
}
\author{
Nils Tjaden \href{mailto:n.b.tjaden@utwente.nl}{n.b.tjaden@utwente.nl}
}
