% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_deviation.R
\name{tune_CORT}
\alias{tune_CORT}
\title{Tuning function (phi) for CORT}
\usage{
tune_CORT(x)
}
\arguments{
\item{x}{a vector, containing the result of CORT()}
}
\value{
a vector, containing the tuned value
}
\description{
Tunes the first order correlation coefficient from [-1,1] to (0,2).
This is based on equation 5 in Gao et al. (2019).
}
\examples{
data("EuStockMarkets")
tune_CORT(CORT(X=EuStockMarkets[ ,1], Y=EuStockMarkets[ ,2]))

plot(tune_CORT(seq(from=-1, to=1, by=.01))~seq(from=-1, to=1, by=.01),
 type="l", xlab="CORT value", ylab="tuned value")

}
\references{
Gao, Y. et al. (2019) ‘Measuring spatio-temporal autocorrelation
in time series data of collective human mobility’, Geo-spatial Information
Science, 22(3), pp. 166–173. doi:10.1080/10095020.2019.1643609.
}
\author{
Nils Tjaden \href{mailto:n.b.tjaden@utwente.nl}{n.b.tjaden@utwente.nl}
}
